Added tests for reading and writing a crossword

+++ palabralib/tests/test_files.py
@@ -18,7 +18,18 @@
 import os
 import unittest
 
-from palabralib.files import read_crossword, write_crossword_to_xml
+from lxml import etree
+
+from palabralib.files import (
+    read_crossword, 
+    write_crossword_to_xml,
+    _read_metadata,
+    _write_metadata,
+    _read_grid,
+    _write_grid,
+    _read_clues,
+    _write_clues,
+)
 from palabralib.grid import Grid
 from palabralib.puzzle import Puzzle
 
@@ -30,19 +41,57 @@ class FilesTestCase(unittest.TestCase):
         self.puzzle.filename = self.LOCATION
         
     def tearDown(self):
-        os.remove(self.LOCATION)
+        if os.path.exists(self.LOCATION):
+            os.remove(self.LOCATION)
         
-    def testReadWrite(self):
+    def testReadWriteCrossword(self):
         clues = {"text": "foo", "explanation": "bar"}
         self.puzzle.grid.set_block(0, 0, True)
         self.puzzle.grid.set_char(1, 1, "A")
         self.puzzle.grid.cell(2, 2)["clues"]["across"] = clues
+        self.puzzle.metadata = {"title": "A", "author": "B"}
     
         write_crossword_to_xml(self.puzzle)
         puzzle = read_crossword(self.LOCATION)
-
         self.assertEqual(puzzle.grid.is_block(0, 0), True)
         self.assertEqual(puzzle.grid.get_char(1, 1), "A")
         self.assertEqual(puzzle.grid.cell(2, 2)["clues"]["across"], clues)
         for x, y in self.puzzle.grid.cells():
             self.assertEqual(puzzle.grid.cell(x, y), self.puzzle.grid.cell(x, y))
+        self.assertEqual(puzzle.metadata["title"], "A")
+        self.assertEqual(puzzle.metadata["author"], "B")
+        
+    def testReadWriteMetadata(self):
+        metadata = {"title": "A"
+            , "author": "B"
+            , "copyright": "C"
+            , "description": "D"}
+        root = etree.Element("root")
+        _write_metadata(root, metadata)
+        result = _read_metadata(root[0])
+        self.assertEqual(result, metadata)
+        
+    def testReadWriteGrid(self):
+        clues = {"text": "C", "explanation": "D"}
+        self.puzzle.grid.set_block(5, 5, True)
+        self.puzzle.grid.set_char(6, 6, "A")
+        root = etree.Element("root")
+        _write_grid(root, self.puzzle.grid)
+        result = _read_grid(root[0])
+        self.assertEqual(result.is_block(5, 5), True)
+        self.assertEqual(result.get_char(6, 6), "A")
+        
+    def testReadWriteClues(self):
+        a = {"text": "foo", "explanation": "bar"}
+        self.puzzle.grid.cell(0, 0)["clues"]["across"] = a
+        d = {"text": "bar", "explanation": "foo"}
+        self.puzzle.grid.cell(0, 0)["clues"]["down"] = d
+        root = etree.Element("root")
+        _write_clues(root, self.puzzle.grid, "across")
+        _write_clues(root, self.puzzle.grid, "down")
+        dir_a, result_a = _read_clues(root[0])
+        dir_d, result_d = _read_clues(root[1])
+        self.assertEqual(dir_a, "across")
+        self.assertEqual(dir_d, "down")
+        self.assertEqual((0, 0, a) in result_a, True)
+        self.assertEqual((0, 0, d) in result_d, True)
